add func to generate flake key (!)
beware! the algorithm is not exactly the one proposed in the paper, but a slight
modification is used. We changed the ordering as follows:

From ji0 -> ij0 -> ij1 -> ji1 -> ji0
To   ji0 -> ij1 -> ij0 -> ji1 -> ji0

This way the ordering is symmetric if you switch the view from user i to user j.

!!! CRYPTOGRAPHIC CONSEQUENCES UNKNOWN !!!