Do not close sockets in signal handlers.
The only valid thing to do in a signal handler is to set a flag OR
write a byte to a pipe which could be used in the select-loop to break upon the
signal.  Any thing other than that will interfere with any pending syscalls and
could lead to undefined behavior.